refactoring website body
