+++ src/client/websites/WebsiteLayoutComponents/Body/WebsiteBody.js
@@ -1,42 +1,24 @@
-import React, { useCallback, useEffect, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { Link, withRouter } from 'react-router-dom';
-import { isEmpty, get, map, debounce, isEqual, reverse } from 'lodash';
+import { isEmpty, get } from 'lodash';
 import { Helmet } from 'react-helmet';
 import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { connect } from 'react-redux';
-import { Map } from 'pigeon-maps';
-import Overlay from 'pigeon-overlay';
-import { getCoordinates } from '../../../../store/userStore/userActions';
 import {
   setFilterFromQuery,
   setMapForSearch,
-  setSearchInBox,
-  setShowSearchResult,
   setWebsiteSearchType,
 } from '../../../../store/searchStore/searchActions';
 import SearchAllResult from '../../../search/SearchAllResult/SearchAllResult';
-import mapProvider from '../../../helpers/mapProvider';
-import { getParsedMap } from '../../../components/Maps/mapHelper';
-import CustomMarker from '../../../components/Maps/CustomMarker';
 import {
   getObjectAvatar,
   getObjectMapInArray,
   getObjectName,
 } from '../../../helpers/wObjectHelper';
-import { handleAddMapCoordinates } from '../../../rewards/rewardsHelper';
-import {
-  getCurrentAppSettings,
-  getReservedCounter,
-  putUserCoordinates,
-} from '../../../../store/appStore/appActions';
-import {
-  getWebsiteObjWithCoordinates,
-  setShowReload,
-} from '../../../../store/websiteStore/websiteActions';
-import { distanceInMBetweenEarthCoordinates, getFirstOffsetNumber } from '../../helper';
-import ObjectOverlayCard from '../../../objectCard/ObjectOverlayCard/ObjectOverlayCard';
+import { getReservedCounter } from '../../../../store/appStore/appActions';
+import { setShowReload } from '../../../../store/websiteStore/websiteActions';
 import {
   getConfigurationValues,
   getHostAddress,
@@ -52,109 +34,24 @@ import {
   getWebsiteSearchType,
   tagsCategoryIsEmpty,
 } from '../../../../store/searchStore/searchSelectors';
-import {
-  getShowReloadButton,
-  getWobjectsPoint,
-} from '../../../../store/websiteStore/websiteSelectors';
+import { getShowReloadButton } from '../../../../store/websiteStore/websiteSelectors';
 import { createFilterBody, parseTagsFilters } from '../../../discoverObjects/helper';
-import MapControllers from '../../../widgets/MapControllers/MapControllers';
-import TagFilters from '../../TagFilters/TagFilters';
+import MainMap from '../../MainMap/MainMap';
 
 import './WebsiteBody.less';
 
 const WebsiteBody = props => {
-  const [boundsParams, setBoundsParams] = useState({
-    topPoint: [],
-    bottomPoint: [],
-  });
-  const [infoboxData, setInfoboxData] = useState(null);
   const [hoveredCardPermlink, setHoveredCardPermlink] = useState('');
-  const [height, setHeight] = useState('100%');
-  const [showLocation, setShowLocation] = useState(false);
-  const [area, setArea] = useState({ center: [], zoom: 11, bounds: [] });
   const reservedButtonClassList = classNames('WebsiteBody__reserved', {
     'WebsiteBody__reserved--withMobileFilters': props.isActiveFilters,
   });
-  let queryCenter = props.query.get('center');
   const isMobile = props.screenSize === 'xsmall' || props.screenSize === 'small';
-  const getCurrentConfig = config =>
-    isMobile ? get(config, 'mobileMap', {}) : get(config, 'desktopMap', {});
   const mapClassList = classNames('WebsiteBody__map', { WebsiteBody__hideMap: props.isShowResult });
-  let mapHeight = 'calc(100vh - 57px)';
-  const mapRef = useRef();
-
-  if (height && isMobile) mapHeight = `${height - 57}px`;
-
-  const getCenter = config => get(getCurrentConfig(config), 'center');
-  const getZoom = config => get(getCurrentConfig(config), 'zoom');
-  const setCurrMapConfig = (center, zoom) => setArea({ center, zoom, bounds: [] });
-
-  if (queryCenter) {
-    queryCenter = queryCenter.split(',').map(item => Number(item));
-  }
-
-  useEffect(() => {
-    if (mapRef.current && props.query.get('showPanel')) {
-      const bounce = mapRef.current.getBounds();
-
-      if (bounce.ne[0] && bounce.sw[0]) {
-        props.setShowSearchResult(true);
-        setBoundsParams({
-          topPoint: [bounce.ne[1], bounce.ne[0]],
-          bottomPoint: [bounce.sw[1], bounce.sw[0]],
-        });
-      }
-    }
-  }, [mapRef.current]);
-
-  const getCoordinatesForMap = async () => {
-    let currZoom;
-    let currCenter;
-
-    if (!isEmpty(queryCenter)) {
-      currZoom = +props.query.get('zoom');
-      currCenter = queryCenter;
-    } else {
-      const currLocation = await props.getCoordinates();
-      const res = await props.getCurrentAppSettings();
-      const siteConfig = get(res, 'configuration');
-
-      currZoom = getZoom(siteConfig) || 6;
-      currCenter = getCenter(siteConfig);
-
-      currCenter = isEmpty(currCenter)
-        ? [get(currLocation, ['value', 'latitude']), get(currLocation, ['value', 'longitude'])]
-        : currCenter;
-    }
-
-    setCurrMapConfig(currCenter, currZoom);
-
-    return { currCenter, currZoom };
-  };
-
-  const handleSetMapForSearch = () => {
-    if (!isEmpty(area.center)) {
-      props.query.set('showPanel', true);
-      props.query.set('center', area.center);
-      props.query.set('zoom', area.zoom);
-      props.setMapForSearch({
-        coordinates: reverse([...area.center]),
-        ...boundsParams,
-      });
-    }
-
-    props.history.push(`?${props.query.toString()}`);
-    localStorage.setItem('query', props.query.toString());
-  };
 
   useEffect(() => {
     const query = props.location.search;
-    const handleResize = () => setHeight(window.innerHeight);
-
-    setHeight(window.innerHeight);
 
     if (props.isAuth) props.getReservedCounter();
-
     if (query) {
       const filterBody = createFilterBody(parseTagsFilters(query));
       const type = props.query.get('type');
@@ -162,60 +59,7 @@ const WebsiteBody = props => {
       if (type) props.setWebsiteSearchType(type);
       if (!isEmpty(filterBody)) props.setFilterFromQuery(filterBody);
     }
-
-    getCoordinatesForMap();
-
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
   }, []);
-
-  useEffect(() => {
-    if (props.isShowResult) {
-      handleSetMapForSearch();
-    } else {
-      props.setMapForSearch({});
-      props.setShowReload(false);
-      props.setSearchInBox(true);
-      props.history.push(`?${props.query.toString()}`);
-    }
-  }, [props.isShowResult]);
-
-  useEffect(() => {
-    const { topPoint, bottomPoint } = boundsParams;
-
-    if (!isEmpty(topPoint) && !isEmpty(bottomPoint))
-      props
-        .getWebsiteObjWithCoordinates(props.searchString, { topPoint, bottomPoint }, 50)
-        .then(res => {
-          if (!isEmpty(props.searchMap)) {
-            const distance = distanceInMBetweenEarthCoordinates(
-              reverse([...props.searchMap.coordinates]),
-              area.center,
-            );
-
-            if (distance > 20 && !props.showReloadButton) props.setShowReload(true);
-            if (!distance) props.setShowReload(false);
-          }
-          if (!isEmpty(queryCenter)) {
-            const { wobjects } = res.value;
-            const queryPermlink = props.query.get('permlink');
-            const currentPoint =
-              get(infoboxData, ['wobject', 'author_permlink']) !== queryPermlink
-                ? wobjects.find(wobj => wobj.author_permlink === queryPermlink)
-                : null;
-
-            if (currentPoint) {
-              setInfoboxData({
-                wobject: currentPoint,
-                coordinates: queryCenter,
-              });
-            }
-          }
-        });
-  }, [props.userLocation, boundsParams, props.query.toString()]);
 
   const aboutObject = get(props, ['configuration', 'aboutObject'], {});
   const configLogo = isMobile ? props.configuration.mobileLogo : props.configuration.desktopLogo;
@@ -225,134 +69,22 @@ const WebsiteBody = props => {
   const objName = getObjectName(aboutObject);
   const title = get(aboutObject, 'title', '') || objName;
 
-  const reloadSearchList = () => {
-    handleSetMapForSearch();
-    props.setShowReload(false);
-  };
-
-  const handleOnBoundsChanged = useCallback(
-    debounce(data => {
-      if (!isEmpty(data) && data.ne[0] && data.sw[0]) {
-        setBoundsParams({
-          topPoint: [data.ne[1], data.ne[0]],
-          bottomPoint: [data.sw[1], data.sw[0]],
-        });
-      }
-    }, 300),
-    [],
-  );
-
-  const onBoundsChanged = ({ center, zoom, bounds }) => {
-    if (!isEmpty(center)) setArea({ center, zoom, bounds });
-    if (!isEqual(bounds, area.bounds)) handleOnBoundsChanged(bounds);
-  };
+  const reloadSearchList = () => props.setShowReload(false);
 
   const handleHoveredCard = permlink => setHoveredCardPermlink(permlink);
 
-  const handleMarkerClick = useCallback(
-    ({ payload, anchor }) => {
-      handleAddMapCoordinates(anchor);
-
-      if (get(infoboxData, 'coordinates', []) === anchor) setInfoboxData({ infoboxData: null });
-
-      props.query.set('center', anchor);
-      props.query.set('zoom', area.zoom);
-      props.query.set('permlink', payload.author_permlink);
-      props.history.push(`?${props.query.toString()}`);
-      setInfoboxData({ wobject: payload, coordinates: anchor });
-    },
-    [area.zoom, props.location.search],
-  );
-
-  const getMarkers = useCallback(
-    wObjects =>
-      !isEmpty(wObjects) &&
-      map(wObjects, wobject => {
-        const parsedMap = getParsedMap(wobject);
-        const latitude = get(parsedMap, ['latitude']);
-        const longitude = get(parsedMap, ['longitude']);
-        const isMarked = Boolean(
-          get(wobject, 'campaigns') || !isEmpty(get(wobject, 'propositions')),
-        );
-
-        return latitude && longitude ? (
-          <CustomMarker
-            key={get(wobject, '_id')}
-            isMarked={isMarked}
-            anchor={[+latitude, +longitude]}
-            payload={wobject}
-            onClick={handleMarkerClick}
-            onDoubleClick={() => setInfoboxData(null)}
-            hoveredWobj={hoveredCardPermlink === wobject.author_permlink}
-          />
-        ) : null;
-      }),
-    [props.wobjectsPoint, hoveredCardPermlink],
-  );
-
-  const getOverlayLayout = useCallback(() => {
-    if (!infoboxData) return null;
-
-    const currentWobj = infoboxData;
-    const name = getObjectName(currentWobj.wobject);
-    const wobject = get(currentWobj, 'wobject', {});
-    const firstOffsetNumber = getFirstOffsetNumber(name);
-    const setQueryInStorage = () => localStorage.setItem('query', props.query);
-
-    return (
-      <Overlay
-        anchor={infoboxData.coordinates}
-        offset={[firstOffsetNumber, 160]}
-        className="WebsiteBody__overlay"
-      >
-        <div className="WebsiteBody__overlay-wrap" role="presentation" onClick={setQueryInStorage}>
-          <ObjectOverlayCard wObject={wobject} showParent={props.searchType !== 'restaurant'} />
-        </div>
-      </Overlay>
-    );
-  }, [infoboxData]);
-
-  const incrementZoom = () => setArea({ ...area, zoom: area.zoom + 1 });
-  const decrementZoom = () => setArea({ ...area, zoom: area.zoom - 1 });
-  const setLocationFromNavigator = position => {
-    const { latitude, longitude } = position.coords;
-
-    setArea({ ...area, center: [latitude, longitude] });
-    setShowLocation(true);
-    props.putUserCoordinates({ latitude, longitude });
-  };
-
-  const setLocationFromApi = () => {
-    setShowLocation(false);
-    setArea({
-      ...area,
-      center: [props.userLocation.lat, props.userLocation.lon],
-    });
-  };
-
-  const handleSetFiltersInUrl = (category, value) => {
-    if (value === 'all') props.query.delete(category);
-    else props.query.set(category, value);
-
-    props.history.push(`?${props.query.toString()}`);
-  };
-
   const handleUrlWithChangeType = type => {
-    let query = `?type=${type}&center=${area.center}&zoom=${area.zoom}`;
+    let query = `?type=${type}`;
 
     if (props.searchString) query = `${query}&searchString=${props.searchString}`;
 
-    setInfoboxData(null);
     props.history.push(query);
   };
 
   const setQueryFromSearchList = obj => {
     const objMap = getObjectMapInArray(obj);
 
-    props.query.set('center', area.center);
-    props.query.set('zoom', area.zoom);
     props.query.set('permlink', obj.author_permlink);
-
     if (objMap) props.query.set('center', objMap);
     if (props.searchString) props.query.set('searchString', props.searchString);
   };
@@ -394,19 +126,18 @@ const WebsiteBody = props => {
         reloadSearchList={reloadSearchList}
         searchType={props.searchType}
         handleHoveredCard={handleHoveredCard}
-        handleSetFiltersInUrl={handleSetFiltersInUrl}
         handleUrlWithChangeType={handleUrlWithChangeType}
         setQueryInLocalStorage={setQueryInLocalStorage}
         setQueryFromSearchList={setQueryFromSearchList}
         deleteShowPanel={deleteShowPanel}
       />
-      <div className={mapClassList} style={{ height: mapHeight }}>
+      <div className={mapClassList}>
         {currentLogo && (
           <Link to={logoLink}>
             <img className="WebsiteBody__logo" src={currentLogo} alt="your logo" />
           </Link>
         )}
-        {!isEmpty(area.center) && !isEmpty(props.configuration) && (
+        {!isEmpty(props.configuration) && (
           <React.Fragment>
             {Boolean(props.counter) && props.isAuth && (
               <Link to="/rewards/reserved" className={reservedButtonClassList}>
@@ -414,42 +145,7 @@ const WebsiteBody = props => {
                 :&nbsp;&nbsp;&nbsp;&nbsp;{props.counter}
               </Link>
             )}
-            <MapControllers
-              className={'WebsiteBodyControl'}
-              decrementZoom={decrementZoom}
-              incrementZoom={incrementZoom}
-              successCallback={setLocationFromNavigator}
-              rejectCallback={setLocationFromApi}
-            />
-            <Map
-              ref={mapRef}
-              center={area.center}
-              height={Number(mapHeight)}
-              zoom={area.zoom}
-              provider={mapProvider}
-              onBoundsChanged={onBoundsChanged}
-              onClick={({ event }) => {
-                if (event.target.classList.value === 'pigeon-overlays') {
-                  setInfoboxData(null);
-                  props.query.delete('center');
-                  props.query.delete('zoom');
-                  props.query.delete('permlink');
-                  props.history.push(`?${props.query.toString()}`);
-                }
-              }}
-              animate
-              zoomSnap
-            >
-              <TagFilters query={props.query} history={props.history} />
-              {getMarkers(props.wobjectsPoint)}
-              {getOverlayLayout()}
-              {showLocation && (
-                <CustomMarker
-                  anchor={[props.userLocation.lat, props.userLocation.lon]}
-                  currLocation
-                />
-              )}
-            </Map>
+            <MainMap query={props.query} hoveredCardPermlink={hoveredCardPermlink} />
           </React.Fragment>
         )}
       </div>
@@ -468,7 +164,6 @@ WebsiteBody.propTypes = {
   history: PropTypes.shape({
     push: PropTypes.func,
   }).isRequired,
-  getCoordinates: PropTypes.func.isRequired,
   userLocation: PropTypes.shape({
     lat: PropTypes.number,
     lon: PropTypes.number,
@@ -476,19 +171,12 @@ WebsiteBody.propTypes = {
   isShowResult: PropTypes.bool.isRequired,
   configuration: PropTypes.shape().isRequired,
   screenSize: PropTypes.string.isRequired,
-  getWebsiteObjWithCoordinates: PropTypes.func.isRequired,
   searchString: PropTypes.string.isRequired,
   host: PropTypes.string.isRequired,
   getReservedCounter: PropTypes.func.isRequired,
-  putUserCoordinates: PropTypes.func.isRequired,
-  setMapForSearch: PropTypes.func.isRequired,
   setShowReload: PropTypes.func.isRequired,
-  setSearchInBox: PropTypes.func.isRequired,
   setFilterFromQuery: PropTypes.func.isRequired,
-  getCurrentAppSettings: PropTypes.func.isRequired,
   setWebsiteSearchType: PropTypes.func.isRequired,
-  setShowSearchResult: PropTypes.func.isRequired,
-  wobjectsPoint: PropTypes.arrayOf(PropTypes.shape({})),
   counter: PropTypes.number.isRequired,
   searchType: PropTypes.string.isRequired,
   isActiveFilters: PropTypes.bool.isRequired,
@@ -505,7 +193,6 @@ WebsiteBody.propTypes = {
 };
 
 WebsiteBody.defaultProps = {
-  wobjectsPoint: [],
   searchString: '',
   isAuth: false,
   showReloadButton: false,
@@ -517,7 +204,6 @@ export default connect(
     isShowResult: getShowSearchResult(state),
     configuration: getConfigurationValues(state),
     screenSize: getScreenSize(state),
-    wobjectsPoint: getWobjectsPoint(state),
     counter: getReserveCounter(state),
     isAuth: getIsAuthenticated(state),
     query: new URLSearchParams(ownProps.location.search),
@@ -529,16 +215,10 @@ export default connect(
     isActiveFilters: tagsCategoryIsEmpty(state),
   }),
   {
-    getCoordinates,
     setWebsiteSearchType,
-    getWebsiteObjWithCoordinates,
     getReservedCounter,
-    putUserCoordinates,
-    getCurrentAppSettings,
     setMapForSearch,
     setShowReload,
-    setSearchInBox,
     setFilterFromQuery,
-    setShowSearchResult,
   },
 )(withRouter(WebsiteBody));
