+++ src/common/routes/switchRoutes.js
@@ -2,7 +2,7 @@ import websitesRoutes from './configs/websitesRoutes';
 import routes from './configs/routes';
 
 const switchRoutes = host => {
-  if (host.includes('dining')) return websitesRoutes(host);
+  if (host.includes('host')) return websitesRoutes(host);
 
   return routes;
 };
